+++ packages/react/src/components/DashboardEditor/DashboardEditor.story.jsx
@@ -177,7 +177,6 @@ export const Default = () => (
       headerEditTitleButton: 'Edit title updated',
     }}
     onAddImage={action('onAddImage')}
-    onEditTitle={action('onEditTitle')}
     onImport={action('onImport')}
     onExport={action('onExport')}
     onDelete={action('onDelete')}
@@ -380,7 +379,6 @@ export const WithInitialValue = () => (
         ],
       },
     }}
-    onEditTitle={action('onEditTitle')}
     onImport={action('onImport')}
     onExport={action('onExport')}
     onDelete={action('onDelete')}
@@ -411,6 +409,12 @@ export const WithInitialValue = () => (
 
 WithInitialValue.story = {
   name: 'with initialValue',
+};
+
+export const WithEditableTitle = () => <DashboardEditor title="Custom dashboard" isTitleEditable />;
+
+WithEditableTitle.story = {
+  namd: 'with editable title',
 };
 
 export const SummaryDashboardWithInitialValue = () => (
@@ -581,7 +585,6 @@ export const SummaryDashboardWithInitialValue = () => (
           ],
         },
       }}
-      onEditTitle={action('onEditTitle')}
       onImport={action('onImport')}
       onExport={action('onExport')}
       onDelete={action('onDelete')}
@@ -651,7 +654,6 @@ export const WithCustomOnCardChange = () => (
       ],
       layouts: {},
     }}
-    onEditTitle={action('onEditTitle')}
     onImport={action('onImport')}
     onExport={action('onExport')}
     onDelete={action('onDelete')}
@@ -691,7 +693,6 @@ export const WithNotifications = () => (
     isSubmitDisabled={boolean('isSubmitDisabled', false)}
     isSubmitLoading={boolean('isSubmitLoading', false)}
     title={text('title', 'My dashboard')}
-    onEditTitle={action('onEditTitle')}
     onImport={action('onImport')}
     onExport={action('onExport')}
     onDelete={action('onDelete')}
@@ -748,7 +749,6 @@ export const WithBreakpointSwitcher = () => (
       isSubmitLoading={boolean('isSubmitLoading', false)}
       title={text('title', 'My dashboard')}
       onAddImage={action('onAddImage')}
-      onEditTitle={action('onEditTitle')}
       onImport={action('onImport')}
       onExport={action('onExport')}
       onDelete={action('onDelete')}
@@ -815,7 +815,6 @@ export const CustomCardPreviewRenderer = () => (
       ],
       layouts: {},
     })}
-    onEditTitle={action('onEditTitle')}
     onImport={action('onImport')}
     onExport={action('onExport')}
     onDelete={action('onDelete')}
@@ -908,7 +907,6 @@ export const I18N = () => (
     dataItems={mockDataItems}
     availableImages={images}
     onAddImage={action('onAddImage')}
-    onEditTitle={action('onEditTitle')}
     onImport={action('onImport')}
     onExport={action('onExport')}
     onDelete={action('onDelete')}
@@ -1349,7 +1347,6 @@ export const WithCustomCards = () => {
       }}
       icons={{ CUSTOM1: <DataScientistIcon />, CUSTOM2: <EmptystateDefaultIcon /> }}
       onAddImage={action('onAddImage')}
-      onEditTitle={action('onEditTitle')}
       onImport={action('onImport')}
       onExport={action('onExport')}
       onDelete={action('onDelete')}
@@ -1393,4 +1390,89 @@ export const WithCustomCards = () => {
 
 WithCustomCards.story = {
   name: 'With custom cards',
+};
+
+export const withGetDefaultCard = () => {
+  const renderDefaultCard = ({ cardContent }) => <h5>{cardContent || 'My custom 1 content'}</h5>;
+
+  return (
+    <DashboardEditor
+      title={text('title', 'My dashboard')}
+      getValidDataItems={() => mockDataItems}
+      dataItems={mockDataItems}
+      availableImages={images}
+      onAddImage={action('onAddImage')}
+      onEditTitle={action('onEditTitle')}
+      onImport={action('onImport')}
+      onExport={action('onExport')}
+      onDelete={action('onDelete')}
+      onCancel={action('onCancel')}
+      onSubmit={action('onSubmit')}
+      onImageDelete={action('onImageDelete')}
+      onLayoutChange={action('onLayoutChange')}
+      isSubmitDisabled={boolean('isSubmitDisabled', false)}
+      isSubmitLoading={boolean('isSubmitLoading', false)}
+      availableDimensions={{
+        deviceid: ['73000', '73001', '73002'],
+        manufacturer: ['rentech', 'GHI Industries'],
+      }}
+      onCardChange={(cardConfig) => {
+        if (cardConfig.type === 'MYCUSTOMCARD') {
+          return {
+            ...cardConfig,
+            content: renderDefaultCard(cardConfig),
+          };
+        }
+        return cardConfig;
+      }}
+      getDefaultCard={(type) => {
+        switch (type) {
+          case 'MYCUSTOMCARD':
+            return {
+              id: 'custom1Id',
+              type: 'MYCUSTOMCARD',
+              title: 'My custom 1',
+              size: 'MEDIUM',
+              content: renderDefaultCard,
+              renderEditContent: (onChange, cardConfig) => [
+                {
+                  header: { title: 'Edit custom 1', tooltip: { tooltipText: 'tooltip' } },
+                  content: (
+                    <TextInput
+                      onChange={(e) => onChange({ ...cardConfig, cardContent: e.target.value })}
+                    />
+                  ),
+                },
+              ],
+            };
+          case 'MYCUSTOMCARD2':
+            return {
+              id: 'custom2Id',
+              type: 'MYCUSTOMCARD2',
+              title: 'My custom 2',
+              size: 'MEDIUMTHIN',
+              content: () => <h5>My custom 2 content</h5>,
+            };
+          default:
+            return {
+              id: 'defaultId',
+              title: 'defaultCard',
+              size: 'MEDIUM',
+              content: () => <h5>Default card</h5>,
+            };
+        }
+      }}
+      supportedCardTypes={array('supportedCardTypes', ['MYCUSTOMCARD', 'MYCUSTOMCARD2'])}
+      breakpointSwitcher={{ enabled: true }}
+      headerBreadcrumbs={[
+        <Link href="www.ibm.com">Dashboard library</Link>,
+        <Link href="www.ibm.com">Favorites</Link>,
+      ]}
+      isLoading={boolean('isLoading', false)}
+    />
+  );
+};
+
+withGetDefaultCard.story = {
+  name: 'With get default card',
 };